+++ packages/core/src/classify.ts
@@ -1,5 +1,5 @@
-import type { Element } from '@yk-yong/react-native-richtext-dom'
-import type { ComputedStyle } from '@yk-yong/react-native-richtext-css'
+import type { Element } from '@quanta-studio/react-native-richtext-dom'
+import type { ComputedStyle } from '@quanta-studio/react-native-richtext-css'
 
 export type Display = ComputedStyle['control']['display']